TASK: Followup #3316 merge `UriHelper`'s
